refactor: table mutation aggregator
